requirements.txt
streamlit
transformers
pygments
google-generativeai
